Move hooks to src root
