+++ src/components/burger-constructor/burger-constructor.jsx
@@ -5,7 +5,7 @@ import { Button, CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-co
 import BurgerConstructorItem from "./burger-constructor-item";
 import OrderDetails from "./order-details";
 import Modal from "../modal/modal";
-import { useModal } from "../hooks/useModal";
+import { useModal } from "../../hooks/useModal";
 import styles from "./burger-constructor.module.css";
 
 function BurgerConstructor(props) {
